Merge remote-tracking branch 'origin/development' into feature/shop_template
Conflicts:
	.meteor/versions
First success, can override templates, added a new package for this. Made a function for uploading an img in header, but now, i did something wrong and it doesnt work anymore, happy debugging!
The error was "inputhandler is not defined", so it didnt show my self created template. It must be an ID-conflict, i have to write my own function to check if the upload/input field/tag has changed to get my click-event on the placeholder-img in header.
The Button for img-upload in header works on every route, now i can't store obj in collections 'cause i can't get the 'media-collection'.
fixed the "header-image-upload"
found bug, if template is replaced, helpers and probably events will only be called if they're defined in the "old template".
Now i can upload an image to mongoDB and it will be shown from DB, getting the img by name.